requirements.txt
easyocr==1.6.2
pandas==1.5.3
Pillow==10.0.0
Requests==2.31.0
streamlit==1.25.0